refactor(cross_model_label): graduate to top-level layout, add docs-site page
Merge `main` and relocate from `experimental/cross_model_label` to top-level
`pydantic_ai_harness/cross_model_label/`, matching #347/#354. Drops the
experimental import warning, fixes import paths, moves tests. Adds
`docs/cross-model-label.md` (nav.json, index.md, parity gate) and a README matrix
row. No behavior change.
