+++ docs/nav.json
@@ -10,6 +10,7 @@
       { "label": "Pydantic AI Docs", "slug": "pydantic-ai-docs" },
       { "label": "Compaction", "slug": "compaction" },
       { "label": "Overflowing Tool Output", "slug": "overflowing-tool-output" },
+      { "label": "Cross-Model Label", "slug": "cross-model-label" },
       { "label": "Step Persistence", "slug": "step-persistence" },
       { "label": "Media", "slug": "media" },
       { "label": "Subagents", "slug": "subagents" },
