+++ tests/test_docs_parity.py
@@ -120,6 +120,7 @@ _CAPABILITY_PAGE_META = {
     'pydantic-ai-docs.md': ('docs', 'Pydantic AI Docs'),
     'compaction.md': ('compaction', 'Compaction'),
     'overflowing-tool-output.md': ('overflowing_tool_output', 'Overflowing Tool Output'),
+    'cross-model-label.md': ('cross_model_label', 'Cross-Model Label'),
     'step-persistence.md': ('step_persistence', 'Step Persistence'),
     'media.md': ('media', 'Media Externalization'),
     'subagents.md': ('subagents', 'Subagents'),
